+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/bankorder/BankOrderLineService.java
@@ -26,6 +26,7 @@ import com.axelor.apps.bankpayment.db.repo.BankOrderRepository;
 import com.axelor.apps.bankpayment.db.repo.EbicsPartnerRepository;
 import com.axelor.apps.bankpayment.exception.IExceptionMessage;
 import com.axelor.apps.base.db.Address;
+import com.axelor.apps.base.db.Bank;
 import com.axelor.apps.base.db.BankDetails;
 import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.Currency;
@@ -34,6 +35,7 @@ import com.axelor.apps.base.db.repo.BankDetailsRepository;
 import com.axelor.apps.base.service.CurrencyService;
 import com.axelor.apps.base.service.PartnerService;
 import com.axelor.apps.tool.StringTool;
+import com.axelor.db.Model;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.i18n.I18n;
@@ -53,14 +55,17 @@ public class BankOrderLineService {
 
   protected BankDetailsRepository bankDetailsRepo;
   protected CurrencyService currencyService;
-  protected PartnerService partnerService;
+  protected BankOrderLineOriginService bankOrderLineOriginService;
 
   @Inject
   public BankOrderLineService(
-      BankDetailsRepository bankDetailsRepo, CurrencyService currencyService) {
+      BankDetailsRepository bankDetailsRepo,
+      CurrencyService currencyService,
+      BankOrderLineOriginService bankOrderLineOriginService) {
 
     this.bankDetailsRepo = bankDetailsRepo;
     this.currencyService = currencyService;
+    this.bankOrderLineOriginService = bankOrderLineOriginService;
   }
 
   /**
@@ -80,7 +85,8 @@ public class BankOrderLineService {
       Currency currency,
       LocalDate bankOrderDate,
       String receiverReference,
-      String receiverLabel)
+      String receiverLabel,
+      Model origin)
       throws AxelorException {
 
     BankDetails receiverBankDetails = bankDetailsRepo.findDefaultByPartner(partner);
@@ -94,7 +100,8 @@ public class BankOrderLineService {
         currency,
         bankOrderDate,
         receiverReference,
-        receiverLabel);
+        receiverLabel,
+        origin);
   }
 
   /**
@@ -114,7 +121,8 @@ public class BankOrderLineService {
       Currency currency,
       LocalDate bankOrderDate,
       String receiverReference,
-      String receiverLabel)
+      String receiverLabel,
+      Model origin)
       throws AxelorException {
 
     return this.createBankOrderLine(
@@ -126,7 +134,8 @@ public class BankOrderLineService {
         currency,
         bankOrderDate,
         receiverReference,
-        receiverLabel);
+        receiverLabel,
+        origin);
   }
 
   /**
@@ -150,7 +159,8 @@ public class BankOrderLineService {
       Currency currency,
       LocalDate bankOrderDate,
       String receiverReference,
-      String receiverLabel)
+      String receiverLabel,
+      Model origin)
       throws AxelorException {
 
     BankOrderLine bankOrderLine = new BankOrderLine();
@@ -171,11 +181,20 @@ public class BankOrderLineService {
     bankOrderLine.setReceiverReference(receiverReference);
     bankOrderLine.setReceiverLabel(receiverLabel);
 
+    if (origin != null) {
+      bankOrderLine.addBankOrderLineOriginListItem(
+          bankOrderLineOriginService.createBankOrderLineOrigin(origin));
+    }
+
     if (bankOrderFileFormat
         .getOrderFileFormatSelect()
         .equals(BankOrderFileFormatRepository.FILE_FORMAT_PAIN_XXX_CFONB320_XCT)) {
       bankOrderLine.setBankOrderEconomicReason(bankOrderFileFormat.getBankOrderEconomicReason());
       bankOrderLine.setReceiverCountry(bankOrderFileFormat.getReceiverCountry());
+      Bank bank = bankDetails.getBank();
+      if (bank != null && bank.getCountry() != null) {
+        bankOrderLine.setReceiverCountry(bank.getCountry());
+      }
       bankOrderLine.setPaymentModeSelect(bankOrderFileFormat.getPaymentModeSelect());
       bankOrderLine.setFeesImputationModeSelect(bankOrderFileFormat.getFeesImputationModeSelect());
       bankOrderLine.setReceiverAddressStr(getReceiverAddress(partner));