+++ axelor-stock/src/main/java/com/axelor/apps/stock/report/IReport.java
@@ -24,7 +24,4 @@ public interface IReport {
   public static final String CONFORMITY_CERTIFICATE = "ConformityCertificate.rptdesign";
   public static final String INVENTORY = "Inventory.rptdesign";
   public static final String STOCK_LOCATION = "StockLocation.rptdesign";
-  public static final String DECLARATION_OF_EXCHANGES_OF_GOODS =
-      "DeclarationOfExchangesOfGoods.rptdesign";
-  public static final String DECLARATION_OF_SERVICES = "DeclarationOfServices.rptdesign";
 }