+++ axelor-stock/src/main/java/com/axelor/apps/stock/report/ITranslation.java
@@ -134,10 +134,4 @@ public interface ITranslation {
   public static final String CONFORMITY_CERTIFICATE_QTY = /*$$(*/ "ConformityCertificate.qty"; /*)*/
   public static final String CONFORMITY_CERTIFICATE_SEQUENCE = /*$$(*/
       "ConformityCertificate.sequence"; /*)*/
-
-  public static final String DECLARATION_OF_EXCHANGES_INTRODUCTION = /*$$(*/ "Introduction"; /*)*/
-  public static final String DECLARATION_OF_EXCHANGES_EXPEDITION = /*$$(*/ "Expedition"; /*)*/
-  public static final String
-      DECLARATION_OF_EXCHANGES_OF_GOODS_BETWEEN_MEMBER_STATES_OF_THE_EUROPEAN_COMMUNITY = /*$$(*/
-          "DECLARATION OF EXCHANGES OF GOODS BETWEEN MEMBER STATES OF THE EUROPEAN COMMUNITY"; /*)*/
 }